+++ baselines/baseline-generator/Program.cs
@@ -5,8 +5,8 @@
 using Microsoft.Azure.Cosmos;
 
 const string EmulatorKey = "C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==";
-var endpoint = "https://localhost:8081";
-string? key = null;
+var endpoint = Environment.GetEnvironmentVariable("AZURE_COSMOS_ENDPOINT") ?? "https://localhost:8081";
+string? key = Environment.GetEnvironmentVariable("AZURE_COSMOS_KEY");
 string? baselineFile = null;
 List<string> queryNames = new List<string>();
 var validateCertificate = true;
