+++ baselines/queries/hybrid/offset_limit_rrf_ft_with_vector.results.json
@@ -1,42 +1,42 @@
 [
   {
-    "index": 51,
-    "title": "Jefferson Township, Berks County, Pennsylvania"
+    "index": 21,
+    "title": "Melissa Gilbert"
   },
   {
-    "index": 54,
-    "title": "Fletcher, Ohio"
+    "index": 37,
+    "title": "Dale Reid"
   },
   {
-    "index": 28,
-    "title": "Lake Miccosukee"
-  },
-  {
-    "index": 70,
-    "title": "Jeff Stover"
-  },
-  {
-    "index": 56,
-    "title": "Newport Tower (Jersey City)"
+    "index": 75,
+    "title": "Vonda Ward"
   },
   {
     "index": 26,
     "title": "Bill Johnson (center)"
   },
   {
+    "index": 35,
+    "title": "Heather McKay"
+  },
+  {
     "index": 24,
     "title": "Thomas W. Osborn"
   },
   {
-    "index": 61,
-    "title": "United States Assistant Attorney General"
+    "index": 87,
+    "title": "Joshua Gregory"
   },
   {
-    "index": 58,
-    "title": "Kjelvatnet (Fauske)"
+    "index": 55,
+    "title": "David E. Davis"
   },
   {
-    "index": 68,
-    "title": "Prince Edward—Hastings"
+    "index": 49,
+    "title": "Perry Greeley Holden"
+  },
+  {
+    "index": 9,
+    "title": "Dennis Gamsy"
   }
 ]
