+++ baselines/queries/hybrid/offset_limit_rrf_unfiltered.results.json
@@ -42,13 +42,5 @@
   {
     "index": 22,
     "title": "Jacob K. Javits Federal Building"
-  },
-  {
-    "index": 6,
-    "title": "Marge vs. the Monorail"
-  },
-  {
-    "index": 73,
-    "title": "173 Hours in Captivity"
   }
 ]
